Consolidate replay probes
